Allow dot entry in float string.
The dot is currently replaced with a comma upon entry. This maintains existing functionality and adds support for automatically converting a dot to a comma.

Add NSNumberFormatter support to Floats

Add tests for Float formatting with a formatter

Add field attributes for number formatting

- positiveFormat: Format of positive numbers
- negativeFormat: Format of negative numbers (defaults to positiveFormat with a - in front)
- groupingSeparator: Separator for whole numbers (e.g., 1,000.00 or 1 000,00)
- decimalSeparator: Separator for decimal place (e.g., 1.23 or 1,23)

Add JSON file with formatted float samples